@@ -1,15 +1,15 @@
 name: weekday_selector_example
 description: A new Flutter project.
-version: 1.0.0+1
-
+version: 1.0.0-nullsafety.0
+publish_to: none
 environment:
-  sdk: ">=2.5.0 <3.0.0"
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   # Used for supporting multiple languages
-  intl: ^0.16.0
+  intl: ^0.17.0-nullsafety.2
   flutter_localizations:
     sdk: flutter
   # The package
