@@ -1,11 +1,11 @@
 name: weekday_selector
 description: A collection of Flutter widgets and classes to help you select weekdays in your apps. Perfect for recurring events, alarms.
-version: 0.4.0
+version: 1.0.0-nullsafety.0
 homepage: https://github.com/smaho-engineering/weekday_selector
 repository: https://github.com/smaho-engineering/weekday_selector
 issue_tracker: https://github.com/smaho-engineering/weekday_selector/issues
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: '>=2.12.0-0 <3.0.0'
 dependencies:
   flutter:
     sdk: flutter
@@ -15,5 +15,5 @@ dev_dependencies:
   # We want to include the package in the tests so that users can understand
   # how the intl package and the weekday selector package work together.
   # The package is also used in the unit tests.
-  intl: ^0.16.1
+  intl: ^0.17.0-nullsafety.2
 flutter:
